Rename agents route to profiles
