+++ src/routes/_layout/skills.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, redirect } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/_layout/skills")({
   beforeLoad: () => {
-    throw redirect({ to: "/v2/agents" })
+    throw redirect({ to: "/v2/profiles" })
   },
   head: () => ({
     meta: [
